# Change order of WebApp initialization during applications checkpoint