--- a/dev/com.ibm.ws.webcontainer.servlet.3.0.factories/bnd.bnd
+++ b/dev/com.ibm.ws.webcontainer.servlet.3.0.factories/bnd.bnd
@@ -73,4 +73,5 @@ Import-Package: \
     com.ibm.ws.managedobject;version=latest,\
     com.ibm.ws.resource;version=latest,\
     com.ibm.ws.session;version=latest, \
-    com.ibm.ws.org.osgi.annotation.versioning;version=latest
+    com.ibm.ws.org.osgi.annotation.versioning;version=latest, \
+	com.ibm.ws.kernel.boot.core;version=latest