--- a/dev/com.ibm.ws.webcontainer.servlet.3.1.factories/src/com/ibm/ws/webcontainer31/osgi/webapp/factory/WebAppFactoryImpl31.java
+++ b/dev/com.ibm.ws.webcontainer.servlet.3.1.factories/src/com/ibm/ws/webcontainer31/osgi/webapp/factory/WebAppFactoryImpl31.java
@@ -10,7 +10,10 @@
  *******************************************************************************/
 package com.ibm.ws.webcontainer31.osgi.webapp.factory;
 
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferenceCardinality;
 
 import com.ibm.websphere.csi.J2EENameFactory;
 import com.ibm.ws.container.service.metadata.MetaDataService;
@@ -21,22 +24,30 @@ import com.ibm.ws.webcontainer.osgi.webapp.WebAppFactory;
 import com.ibm.ws.webcontainer31.osgi.webapp.WebApp31;
 import com.ibm.wsspi.injectionengine.ReferenceContext;
 
+import io.openliberty.checkpoint.spi.CheckpointPhase;
+
 /**
  *
  */
 @Component(service = WebAppFactory.class, property = { "service.vendor=IBM" })
 public class WebAppFactoryImpl31 implements WebAppFactory {
+    private final CheckpointPhase checkpointPhase;
+
+    @Activate
+    public WebAppFactoryImpl31(@Reference(cardinality = ReferenceCardinality.OPTIONAL) CheckpointPhase checkpointPhase) {
+        this.checkpointPhase = checkpointPhase;
+    }
 
     /*
      * (non-Javadoc)
-     * 
+     *
      * @see com.ibm.ws.webcontainer.osgi.webapp.WebAppFactory#createWebApp(com.ibm.ws.webcontainer.osgi.webapp.WebAppConfiguration, java.lang.ClassLoader,
      * com.ibm.wsspi.injectionengine.ReferenceContext, com.ibm.ws.container.service.metadata.MetaDataService, com.ibm.websphere.csi.J2EENameFactory)
      */
     @Override
     public WebApp createWebApp(WebAppConfiguration webAppConfig, ClassLoader moduleLoader, ReferenceContext referenceContext, MetaDataService metaDataService,
                                J2EENameFactory j2eeNameFactory, ManagedObjectService managedObjectService) {
-        return new WebApp31(webAppConfig, moduleLoader, referenceContext, metaDataService, j2eeNameFactory, managedObjectService);
+        return new WebApp31(webAppConfig, moduleLoader, referenceContext, metaDataService, j2eeNameFactory, managedObjectService, checkpointPhase);
     }
 
 }