--- a/dev/com.ibm.ws.webcontainer.servlet.3.1/bnd.bnd
+++ b/dev/com.ibm.ws.webcontainer.servlet.3.1/bnd.bnd
@@ -110,7 +110,8 @@ jakartaeeMe: true
     com.ibm.ws.session;version=latest,\
     com.ibm.ws.org.apache.commons.fileupload;version=latest,\
     com.ibm.ws.anno;version=latest, \
-    com.ibm.ws.org.osgi.annotation.versioning;version=latest
+    com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
+	com.ibm.ws.kernel.boot.core;version=latest
 
 -testpath: \
     org.hamcrest:hamcrest-all;version=1.3, \