--- a/dev/com.ibm.ws.webcontainer.servlet.3.1/test/com/ibm/ws/webcontainer/webapp/MethodsForProgrammaticConfigurationOfServletsTest.java
+++ b/dev/com.ibm.ws.webcontainer.servlet.3.1/test/com/ibm/ws/webcontainer/webapp/MethodsForProgrammaticConfigurationOfServletsTest.java
@@ -61,7 +61,7 @@ public class MethodsForProgrammaticConfigurationOfServletsTest {
         final J2EENameFactory jnf = context.mock(J2EENameFactory.class);
 
         WebAppConfiguration webAppConfig = new WebAppConfiguration(null, "name");
-        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null);
+        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null, null);
 
         webApp.withinContextInitOfProgAddListener = true;
 
@@ -83,7 +83,7 @@ public class MethodsForProgrammaticConfigurationOfServletsTest {
         final J2EENameFactory jnf = context.mock(J2EENameFactory.class);
 
         WebAppConfiguration webAppConfig = new WebAppConfiguration(null, "name");
-        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null);
+        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null, null);
 
         webApp.withinContextInitOfProgAddListener = true;
 
@@ -105,7 +105,7 @@ public class MethodsForProgrammaticConfigurationOfServletsTest {
         final J2EENameFactory jnf = context.mock(J2EENameFactory.class);
 
         WebAppConfiguration webAppConfig = new WebAppConfiguration(null, "name");
-        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null);
+        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null, null);
 
         webApp.withinContextInitOfProgAddListener = true;
 
@@ -127,7 +127,7 @@ public class MethodsForProgrammaticConfigurationOfServletsTest {
         final J2EENameFactory jnf = context.mock(J2EENameFactory.class);
 
         WebAppConfiguration webAppConfig = new WebAppConfiguration(null, "name");
-        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null);
+        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null, null);
 
         webApp.withinContextInitOfProgAddListener = true;
 
@@ -149,7 +149,7 @@ public class MethodsForProgrammaticConfigurationOfServletsTest {
         final J2EENameFactory jnf = context.mock(J2EENameFactory.class);
 
         WebAppConfiguration webAppConfig = new WebAppConfiguration(null, "name");
-        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null);
+        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null, null);
 
         webApp.withinContextInitOfProgAddListener = true;
 
@@ -171,7 +171,7 @@ public class MethodsForProgrammaticConfigurationOfServletsTest {
         final J2EENameFactory jnf = context.mock(J2EENameFactory.class);
 
         WebAppConfiguration webAppConfig = new WebAppConfiguration(null, "name");
-        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null);
+        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null, null);
 
         webApp.withinContextInitOfProgAddListener = true;
 
@@ -193,7 +193,7 @@ public class MethodsForProgrammaticConfigurationOfServletsTest {
         final J2EENameFactory jnf = context.mock(J2EENameFactory.class);
 
         WebAppConfiguration webAppConfig = new WebAppConfiguration(null, "name");
-        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null);
+        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null, null);
 
         webApp.withinContextInitOfProgAddListener = true;
 
@@ -215,7 +215,7 @@ public class MethodsForProgrammaticConfigurationOfServletsTest {
         final J2EENameFactory jnf = context.mock(J2EENameFactory.class);
 
         WebAppConfiguration webAppConfig = new WebAppConfiguration(null, "name");
-        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null);
+        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null, null);
 
         webApp.withinContextInitOfProgAddListener = true;
 
@@ -237,7 +237,7 @@ public class MethodsForProgrammaticConfigurationOfServletsTest {
         final J2EENameFactory jnf = context.mock(J2EENameFactory.class);
 
         WebAppConfiguration webAppConfig = new WebAppConfiguration(null, "name");
-        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null);
+        WebApp31 webApp = new WebApp31(webAppConfig, null, null, mds, jnf, null, null);
 
         webApp.withinContextInitOfProgAddListener = true;