--- a/dev/com.ibm.ws.webcontainer.servlet.4.0/bnd.bnd
+++ b/dev/com.ibm.ws.webcontainer.servlet.4.0/bnd.bnd
@@ -73,7 +73,8 @@ Import-Package: \
     com.ibm.ws.org.apache.commons.fileupload;version=latest,\
     com.ibm.ws.session;version=latest,\
     com.ibm.ws.javaee.dd.common;version=latest,\
-    com.ibm.ws.javaee.dd;version=latest
+    com.ibm.ws.javaee.dd;version=latest,\
+	com.ibm.ws.kernel.boot.core;version=latest
 
 -testpath: \
     org.hamcrest:hamcrest-all;version=1.3, \