Checkpoint restore testing uncovered a large performance gain in first request time by changing the order of these method calls in com.ibm.ws.webcontainer.webapp.WebApp.initialize:

           if (moduleConfig instanceof com.ibm.ws.webcontainer.osgi.container.DeployedModule) {
                // complete the notification here for app manager
                ((com.ibm.ws.webcontainer.osgi.container.DeployedModule) moduleConfig).initTaskComplete();

                if (com.ibm.ejs.ras.TraceComponent.isAnyTracingEnabled() && logger.isLoggable(Level.FINE)) {
                    logger.logp(Level.FINE, CLASS_NAME, "initialize", "set future done for deployedModule -->" + moduleConfig);
                }
            }

                commonInitializationFinally(extensionFactories); // NEVER INVOKED BY

Calling commonInitalizationFinally before moduleConfig.initTaskComplete during an applications checkpoint lowers the restore time to first request from 255ms to 131ms for pingperf and 356ms to 215ms for rest crud.