@tjwatson 2022-03-04
Be sure to run the `io.openliberty.checkpoint_fat` locally on a host that supports criu.